% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_risks.R
\name{get_risks}
\alias{get_risks}
\title{Calculate risks from groupings, scores and overall weightings}
\usage{
get_risks(
  groupings,
  scores,
  weightings = rep(1/length(groupings), length(groupings))
)
}
\description{
Calculate risks from groupings, scores and overall weightings
}
\author{
Finlay Campbell
}
