% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read scores and groupings from data file provided}
\usage{
read_data(file)
}
\description{
Read scores and groupings from data file provided
}
\author{
Finlay Campbell
}
