% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shape.R
\name{read_shape}
\alias{read_shape}
\title{Read shapefiles and check scores have a matching row}
\usage{
read_shape(file, scores)
}
\description{
Read shapefiles and check scores have a matching row
}
\author{
Finlay Campbell
}
