% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskassess-package.R
\docType{package}
\name{riskassess-package}
\alias{riskassess}
\alias{riskassess-package}
\title{riskassess: Shiny App for WHO Risk Assessment}
\description{
This package runs a Shiny app for calibrating risk assessments.
}
\author{
\strong{Maintainer}: Finlay Campbell \email{campellf@who.int}

}
\keyword{internal}
