% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_riskassess.R
\name{run_riskassess}
\alias{run_riskassess}
\title{This function launches the risk assessment Shiny app .}
\usage{
run_riskassess()
}
\description{
This function launches the risk assessment Shiny app .
}
\author{
Finlay Campbell
}
