% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_risk_table.R
\name{vis_risk_table}
\alias{vis_risk_table}
\title{Visualise risk scores in a table.}
\usage{
vis_risk_table(risk, weightings)
}
\arguments{
\item{risk}{The updated risk table access via \code{values$}}

\item{weightings}{The updated weightings access via \code{values$}}
}
\description{
Visualise risk scores in a table.
}
\author{
Finlay Campbell
}
