% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_scores.R
\name{vis_scores}
\alias{vis_scores}
\title{Visualise scores on a map.}
\usage{
vis_scores(
  risks,
  shape,
  value,
  value_label = "Score",
  label_place = FALSE,
  title = NULL
)
}
\description{
Visualise scores on a map.
}
\author{
Finlay Campbell
}
