% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_risks.R
\name{get_risks}
\alias{get_risks}
\title{Calculate risks from groupings, scores and overall weightings}
\usage{
get_risks(
  groupings,
  scores,
  weightings = rep(1/length(groupings), length(groupings))
)
}
\arguments{
\item{groupings}{Named list of named numeric vectors grouping the different
scores into pillars and indicating their relative weighting within-pillar,
as returned by \code{read_data}.}

\item{scores}{data.frame of risk scores for each geographic region, as
returned by \code{read_data}.}

\item{weightings}{Numeric vector of the same length as \code{groupings}
indicating the relative weightings of the different pillars.}
}
\description{
Calculate risks from groupings, scores and overall weightings
}
\author{
Finlay Campbell
}
